==> News_Articles/in_tibet_website/tibetcm/1_all_categories.ipynb <==
from bs4 import BeautifulSoup
import requests
import re


url = "https://www.tibetcm.com/"

headers = {
        'User-Agent': 'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/91.0.4472.124 Safari/537.36'
    }
response = requests.get(url, headers=headers)
response.raise_for_status()

# Parse the page content with BeautifulSoup
soup1 = BeautifulSoup(response.content, 'html.parser')

soup = soup1.find("div", id="nav")
# Find all 'a' tags
links = soup.find_all('a')
all_categories = []
# Extract and print the href attribute from each link
for link in links:
    href = link.get('href')
    if href:
        text = link.get_text()
        all_categories.append((text, href))
        # print(href)

all_categories

def get_page_no(url):
    """
    
    """

    headers = {
        'User-Agent': 'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/91.0.4472.124 Safari/537.36'
    }
    try:
        # Make the request to the URL
        response = requests.get(url, headers=headers)
        response.raise_for_status()
        
        # Parse the page content with BeautifulSoup
        soup = BeautifulSoup(response.content, 'html.parser')
        page = soup.find("div", class_="epages")
        # print("page")
        if page:
            all_links = page.find_all("a")
            if all_links:
                link = all_links[-1].get('href')
                # print("all_links")
                match = re.search(r'index_(\d+)\.html', link)
                return [link, match.group(1)]
            return ["no link last page link" , 1]
        else:
            return ["no link last page link" , 1]

        

    except Exception as e:
        print(e)



## extract page

final_link = []
i = 0
for each_link in all_categories:
    link = each_link[1]
    page_no = get_page_no(link)
    clean_list = []
    clean_list.extend(each_link)
    clean_list.extend(page_no)
    # print(clean_list)
    final_link.append(clean_list)
    i = i+1
    if i == 5:
        print(i)
        # break
final_link

import pandas as pd

df = pd.DataFrame(final_link, columns=['name_of_cate', 'original_url', 'Last_page_link', 'list_page_no'])
df.head()

df.to_csv("./data/All_cate.csv", index=False)

df.shape
